scripts/requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
python-whois>=0.8.0
dnspython>=2.2.0
urllib3>=1.26.0
fpdf2>=2.5.0
